requirements.txt
quart
discordoauth.py
pycryptodomex